+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/north/exception/ExceptionFormat.java
@@ -1,0 +1,14 @@
+package com.asc.loanservice.domain.loan.application.north.exception;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+
+@Builder
+@Getter
+class ExceptionFormat {
+
+    private final String message;
+    private final String description;
+}
