+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/south/LoanEvaluationResultAdapterMapper.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.adapter;
+package com.asc.loanservice.domain.loan.application.south;
 
 import com.asc.loanservice.domain.loan.evaluation.dto.LoanEvaluationResult;
 import org.mapstruct.Mapper;
@@ -8,6 +8,5 @@ import org.springframework.stereotype.Component;
 @Component
 public interface LoanEvaluationResultAdapterMapper {
 
-    com.asc.loanservice.domain.loan.application.south.
-            LoanEvaluationResult mapResult(LoanEvaluationResult loanEvaluationResult);
+    com.asc.loanservice.domain.loan.application.LoanEvaluationResult mapResult(LoanEvaluationResult loanEvaluationResult);
 }
