+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/evaluation/DebtCustomerResult.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.domain.loan.evaluation.south;
+package com.asc.loanservice.domain.loan.evaluation;
 
 
 public enum DebtCustomerResult {
