+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/evaluation/RuleCustomerAge.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.domain.loan.evaluation.rule;
+package com.asc.loanservice.domain.loan.evaluation;
 
 import com.asc.loanservice.domain.loan.application.contract.LoanApplicationRequest;
 import org.springframework.stereotype.Component;
@@ -9,7 +9,7 @@ import java.time.Period;
 import static com.asc.loanservice.constant.Constants.*;
 
 @Component
-class CustomerAgeRule implements Rule {
+class RuleCustomerAge implements Rule {
 
     @Override
     public Boolean isValid(LoanApplicationRequest loanApplicationRequest) {
