+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/evaluation/RuleMaximumInstallmentAmount.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.domain.loan.evaluation.rule;
+package com.asc.loanservice.domain.loan.evaluation;
 
 import com.asc.loanservice.domain.loan.application.contract.LoanApplicationRequest;
 import org.springframework.stereotype.Component;
@@ -9,7 +9,7 @@ import java.math.RoundingMode;
 import static com.asc.loanservice.constant.Constants.*;
 
 @Component
-class MaximumInstallmentAmountRule implements Rule {
+class RuleMaximumInstallmentAmount implements Rule {
 
     @Override
     public Boolean isValid(LoanApplicationRequest loanApplicationRequest) {
