--- src/loan-service/src/main/java/com/asc/loanservice/exception/ExceptionFormat.java
@@ -1,14 +1,0 @@
-package com.asc.loanservice.exception;
-
-
-import lombok.Builder;
-import lombok.Getter;
-
-
-@Builder
-@Getter
-class ExceptionFormat {
-
-    private String message;
-    private String description;
-}
